can: esd_usb2: cancel urb on -EPIPE and -EPROTO
commit 7a31ced3de06e9878e4f9c3abe8f87d9344d8144 upstream. 

In mcba_usb, we have observed that when you unplug the device, the driver will endlessly resubmit failing URBs, which can cause CPU stalls. This issue is fixed in mcba_usb by catching the codes seen on device disconnect (-EPIPE and -EPROTO). 

This driver also resubmits in the case of -EPIPE and -EPROTO, so fix it in the same way. 

Signed-off-by: Martin Kelly <mkelly@xevo.com> 
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de> 
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>