libata: drop WARN from protocol error in ata_sff_qc_issue()
[ Upstream commit 0580b762a4d6b70817476b90042813f8573283fa ] 

ata_sff_qc_issue() expects upper layers to never issue commands on a command protocol that it doesn't implement. While the assumption holds fine with the usual IO path, nothing filters based on the command protocol in the passthrough path (which was added later), allowing the warning to be tripped with a passthrough command with the right (well, wrong) protocol. Failing with AC_ERR_SYSTEM is the right thing to do anyway. 
Remove the unnecessary WARN. 

Reported-by: Dmitry Vyukov <dvyukov@google.com> 
Link: http://lkml.kernel.org/r/CACT4Y+bXkvevNZU8uP6X0QVqsj6wNoUA_1exfTSOzc+SmUtMOA@mail.gmail.com 
Signed-off-by: Tejun Heo <tj@kernel.org> 
Signed-off-by: Sasha Levin <alexander.levin@verizon.com> 
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>