route: also update fnhe_genid when updating a route cache
[ Upstream commit cebe84c6190d741045a322f5343f717139993c08 ] 

Now when ip route flush cache and it turn out all fnhe_genid != genid. If a redirect/pmtu icmp packet comes and the old fnhe is found and all it's members but fnhe_genid will be updated. Then next time when it looks up route and tries to rebind this fnhe to the new dst, the fnhe will be flushed due to fnhe_genid != genid. It causes this redirect/pmtu icmp packet acutally not to be applied. This patch is to also reset fnhe_genid when updating a route cache. 

Fixes: 5aad1de5ea2c ("ipv4: use separate genid for next hop exceptions") 
Acked-by: Hannes Frederic Sowa <hannes@stressinduktion.org> 
Signed-off-by: Xin Long <lucien.xin@gmail.com> 
Signed-off-by: David S. Miller <davem@davemloft.net> 
Signed-off-by: Sasha Levin <alexander.levin@verizon.com> 
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>