route: update fnhe_expires for redirect when the fnhe exists
[ Upstream commit e39d5246111399dbc6e11cd39fd8580191b86c47 ] 

Now when creating fnhe for redirect, it sets fnhe_expires for this new route cache. But when updating the exist one, it doesn't do it. It will cause this fnhe never to be expired. Paolo already noticed it before, in Jianlin's test case, it became even worse: When ip route flush cache, the old fnhe is not to be removed, but only clean it's members. When redirect comes again, this fnhe will be found and updated, but never be expired due to fnhe_expires not being set. So fix it by simply updating fnhe_expires even it's for redirect. Fixes: aee06da6726d ("ipv4: use seqlock for nh_exceptions") 

Reported-by: Jianlin Shi <jishi@redhat.com> 
Acked-by: Hannes Frederic Sowa <hannes@stressinduktion.org> 
Signed-off-by: Xin Long <lucien.xin@gmail.com> 
Signed-off-by: David S. Miller <davem@davemloft.net> 
Signed-off-by: Sasha Levin <alexander.levin@verizon.com> 
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>